Apply feedback
