+++ front/lib/api/wrappers.ts
@@ -1,7 +1,12 @@
 import { DustUserEmailHeader } from "@dust-tt/client";
-import type { WithAPIErrorResponse } from "@dust-tt/types";
+import type {
+  UserTypeWithWorkspaces,
+  WithAPIErrorResponse,
+} from "@dust-tt/types";
 import type { NextApiRequest, NextApiResponse } from "next";
 
+import { getUserFromAuth0Token } from "@app/lib/api/auth0";
+import { getUserWithWorkspaces } from "@app/lib/api/user";
 import {
   Authenticator,
   getAPIKey,
@@ -288,5 +293,63 @@ export function withPublicAPIAuthentication<T, U extends boolean>(
       );
     },
     isStreaming
+  );
+}
+
+/**
+ * This function is a wrapper for Public API routes that require authentication without a workspace with a token from Auth0.
+ */
+export function withAuth0TokenAuthentication<T>(
+  handler: (
+    req: NextApiRequest,
+    res: NextApiResponse<WithAPIErrorResponse<T>>,
+    user: UserTypeWithWorkspaces
+  ) => Promise<void> | void
+) {
+  return withLogging(
+    async (
+      req: NextApiRequest,
+      res: NextApiResponse<WithAPIErrorResponse<T>>
+    ) => {
+      const bearerTokenRes = await getBearerToken(req);
+      if (bearerTokenRes.isErr()) {
+        return apiError(req, res, {
+          status_code: 401,
+          api_error: {
+            type: "not_authenticated",
+            message:
+              "The request does not have valid authentication credentials.",
+          },
+        });
+      }
+      const bearerToken = bearerTokenRes.value;
+      const authMethod = getAuthType(bearerToken);
+
+      if (authMethod !== "access_token") {
+        return apiError(req, res, {
+          status_code: 401,
+          api_error: {
+            type: "not_authenticated",
+            message:
+              "The request does not have valid authentication credentials.",
+          },
+        });
+      }
+
+      const user = await getUserFromAuth0Token(bearerToken);
+      if (!user) {
+        return apiError(req, res, {
+          status_code: 404,
+          api_error: {
+            type: "user_not_found",
+            message: "Could not find the user.",
+          },
+        });
+      }
+
+      const userWithWorkspaces = await getUserWithWorkspaces(user);
+
+      return handler(req, res, userWithWorkspaces);
+    }
   );
 }
