+++ front/pages/api/v1/me.ts
@@ -4,9 +4,7 @@ import type {
 } from "@dust-tt/types";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { getUserFromAuth0Token } from "@app/lib/api/auth0";
-import { getUserWithWorkspaces } from "@app/lib/api/user";
-import { getAuthType, getBearerToken } from "@app/lib/auth";
+import { withAuth0TokenAuthentication } from "@app/lib/api/wrappers";
 import { apiError } from "@app/logger/withlogging";
 
 export type MeResponseBody = {
@@ -21,49 +19,12 @@ export type MeResponseBody = {
 
 async function handler(
   req: NextApiRequest,
-  res: NextApiResponse<WithAPIErrorResponse<MeResponseBody>>
+  res: NextApiResponse<WithAPIErrorResponse<MeResponseBody>>,
+  user: UserTypeWithWorkspaces
 ): Promise<void> {
   switch (req.method) {
     case "GET":
-      const bearerTokenRes = await getBearerToken(req);
-      if (bearerTokenRes.isErr()) {
-        return apiError(req, res, {
-          status_code: 401,
-          api_error: {
-            type: "not_authenticated",
-            message:
-              "The request does not have valid authentication credentials.",
-          },
-        });
-      }
-      const bearerToken = bearerTokenRes.value;
-      const authMethod = getAuthType(bearerToken);
-
-      if (authMethod !== "access_token") {
-        return apiError(req, res, {
-          status_code: 401,
-          api_error: {
-            type: "not_authenticated",
-            message:
-              "The request does not have valid authentication credentials.",
-          },
-        });
-      }
-
-      const user = await getUserFromAuth0Token(bearerToken);
-      if (!user) {
-        return apiError(req, res, {
-          status_code: 404,
-          api_error: {
-            type: "user_not_found",
-            message: "Could not find the user.",
-          },
-        });
-      }
-
-      const userWithWorkspaces = await getUserWithWorkspaces(user);
-
-      return res.status(200).json({ user: userWithWorkspaces });
+      return res.status(200).json({ user });
 
     default:
       return apiError(req, res, {
@@ -76,5 +37,4 @@ async function handler(
   }
 }
 
-// This route is specific, not using the generic wrapper for public API routes.
-export default handler;
+export default withAuth0TokenAuthentication(handler);
